Exploiting C++11 for Iterator interface. Moved Circuit[.hpp] to QICircuit[.h].
